Avoid eager get_slice backend imports
